SA-5 update - COS and OCP
Control is not applicable to either product. Parts a and b do have RH documentation that supports compliance effort of customer. Parts c,d,e are all organization policy and procedure dealing with obtaining, storing, and distributing documents within the org. Definitely not applicable to the products.
Removed trailing   whitespace