AU-11: log retention is provided by offloading logs off-cluster
The control says:
The organization retains audit records for [Assignment:
organization-defined time period consistent with records retention
policy] to provide support for after-the-fact investigations of security
incidents and to meet regulatory and organizational information
retention requirements.

This is not really possible with the retention options that the API
servers themselves offer, so the logs must be forwarder off the cluster
for retention.