MA-4: Just like regular auth
MA-4(c) says:
Employs strong authenticators in the establishment of nonlocal
maintenance and diagnostic sessions

MA-4(e) says:
Terminates session and network connections when nonlocal maintenance is
completed.

Now the way I read the control is that it is meant more for devices or
appliances where you can expect the vendor to log in remotely. In OCP,
there is no concept of a maintenance access, so the best we can do which
IMO covers the control is to require the use of an IDP and remove the
kubeadmin account so that everything is properly audited and tied to
IDP identities.