IA-2(5): No group authenticators except for kubeadmin, let's remove kubeadmin
The control says:
The organization requires individuals to be authenticated with an
individual authenticator when a group authenticator is employed.

The best way to ensure the control is met is through removing kubeadmin
and setting up an IDP, then all user identities can be tied into OCP
identities and no group authentication is performed at all.